moveList增加对文件是否存在的判断，避免中断

+++ src/index.js
@@ -432,7 +432,10 @@ module.exports = function(env, port, hbconfig, wpconfig) {
                     }
                 })
                 var moveItemName = pathArr[pathArr.length - 1];
-                fs.copySync(moveItem, path.join(conf.output, moveItemName));
+                if (fs.existsSync(moveItem)) {
+                    fs.copySync(moveItem, path.join(conf.output, moveItemName));
+                }
+
             })
         }
